chore: Phase 1 & 2 — git hygiene + complete lib domain refactor
Phase 1 — Git hygiene
- Add PWA build artifacts to .gitignore (public/workbox-*.js, sw.js.map,
  swe-worker-*.js.map) — regenerated on every build, must not be tracked
- Add .jules/ AI tooling dir to .gitignore
- git rm --cached all 5 affected files

Phase 2 — Complete lib domain grouping (continues 4f9357a)
- Redirect 13 import statements across 11 files to canonical domain paths:
  @/lib/kemenag-api      → @/lib/quran/kemenag-api      (4 importers)
  @/lib/tafsir-api       → @/lib/quran/tafsir-api       (4 importers incl. test)
  @/lib/quran-segments-api → @/lib/quran/quran-segments-api (2 importers)
  @/lib/analytics-utils  → @/lib/analytics/analytics-utils (3 importers)
  @/lib/activity-tracker → @/lib/analytics/activity-tracker (1 importer)
- Delete 6 stale root-level files:
  kemenag-api.ts, tafsir-api.ts, quran-segments-api.ts,
  analytics-utils.ts, activity-tracker.ts, analytics.ts (barrel re-export)
- Resolves 4-file and 5-file circular import cycles in Quran module

Verified: npx tsc --noEmit → 0 errors
