--- src/lib/activity-tracker.ts
@@ -1,1 +1,0 @@
-export * from './analytics/activity-tracker';
