--- src/lib/analytics.ts
@@ -1,1 +1,0 @@
-export * from './analytics/analytics';
