--- src/lib/kemenag-api.ts
@@ -1,1 +1,0 @@
-export * from './quran/kemenag-api';
