--- src/lib/quran-segments-api.ts
@@ -1,1 +1,0 @@
-export * from './quran/quran-segments-api';
