--- src/lib/tafsir-api.ts
@@ -1,1 +1,0 @@
-export * from './quran/tafsir-api';
